Add basic arithmetic tests for __(u)intcap_t.
Also add __(u)int64_t as a cross check.

The __intcap_t test is not hooked up because it crashes to compiler.

The __uintcap_t test has a number of failures under defined(INCLUDE_XFAIL)
including one case (~var) that crashes the compiler and takes many minutes
to do so.