move trap check to the "right" place
The fault we should expect occurs when we attempt to load a capability
from beyond the end of the va_list.  Even with optimization disabled,
the compiler doesn't actually emit the load until the assert so I've
placed the check there.

Another fault may occur when we pretend the load succeeded and use
whatever trash is in the register to try to load a byte (I found a
sentry in one interation).  I suspect, but have not confirmed, that the
load succeeded by accident when the test was written.